fix(pops-mcp): address CodeRabbit review — hardening, DRY, tests refactor
- Dockerfile: run runtime container as non-root (USER node, --chown=node:node)
- client.ts: throw on missing POPS_API_KEY instead of defaulting to empty string
- index.ts: remove ../../.env load, add 1mb payload limit, export createMcpServer and app for testability; skip listen() in test env
- tools: extract ok()/toolError() to shared utils.ts, removing duplication across all tool files
- inventory.ts, cerebrum.ts: validate required ID/query args and return isError instead of coercing undefined via String()
- tests: move from src/__tests__/ to co-located *.test.ts files per repo convention; add dispatcher-level tests for ListTools/CallTool handlers; add getClient() unit tests (memoization, env wiring, missing key)
- docs: add epic dependency/parallelization sections, fix code block language tags, remove file paths from acceptance criteria, fix us-08 status Pending→Done, update roadmap CI publish note
