+++ apps/pops-mcp/src/tools/index.test.ts
@@ -1,0 +1,26 @@
+import { describe, expect, it, vi } from 'vitest';
+
+import { mockClient } from './test-helpers.js';
+
+vi.mock('../client.js', () => ({ getClient: () => mockClient }));
+
+const { allTools } = await import('./index.js');
+
+describe('allTools', () => {
+  it('exports exactly 14 tools', () => {
+    expect(allTools).toHaveLength(14);
+  });
+
+  it('all tool names are unique', () => {
+    const names = allTools.map((t) => t.name);
+    expect(new Set(names).size).toBe(names.length);
+  });
+
+  it('all tools have a description, inputSchema, and handler', () => {
+    for (const tool of allTools) {
+      expect(tool.description.length).toBeGreaterThan(10);
+      expect(tool.inputSchema['type']).toBe('object');
+      expect(typeof tool.handler).toBe('function');
+    }
+  });
+});
