+++ apps/pops-mcp/src/tools/media.ts
@@ -1,12 +1,7 @@
 import { getClient } from '../client.js';
-
-import type { CallToolResult } from '@modelcontextprotocol/sdk/types.js';
+import { ok } from './utils.js';
 
 import type { ToolDef } from './index.js';
-
-function ok(data: unknown): CallToolResult {
-  return { content: [{ type: 'text', text: JSON.stringify(data, null, 2) }] };
-}
 
 const libraryList: ToolDef = {
   name: 'media.library.list',
