+++ apps/pops-mcp/src/tools/utils.ts
@@ -1,0 +1,9 @@
+import type { CallToolResult } from '@modelcontextprotocol/sdk/types.js';
+
+export function ok(data: unknown): CallToolResult {
+  return { content: [{ type: 'text', text: JSON.stringify(data, null, 2) }] };
+}
+
+export function toolError(message: string): CallToolResult {
+  return { content: [{ type: 'text', text: message }], isError: true };
+}
