+++ apps/pops-mcp/tsconfig.json
@@ -10,5 +10,5 @@
     "sourceMap": true
   },
   "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist"]
+  "exclude": ["node_modules", "dist", "src/**/*.test.ts", "src/**/test-helpers.ts"]
 }
