+++ docs/roadmap.md
@@ -34,12 +34,12 @@ Live status of every theme and epic. Updated as work completes.
 
 > Application-side platform: CI/CD, image packaging + GHCR contract, database operations, cortex runtime. Host-side concerns (ansible, vault, networking, backups, monitoring, Watchtower) are the deployer's responsibility; the knoxio home lab implements them in [`knoxio/homelab-infra`](https://github.com/knoxio/homelab-infra).
 
-| Epic                                   | Status      | Notes                                                                                                |
-| -------------------------------------- | ----------- | ---------------------------------------------------------------------------------------------------- |
-| CI/CD workflows (PRD-016)              | Done        | Per-area quality + docker-build + publish-images.yml; no deploy step in this repo                    |
-| Application packaging & GHCR (PRD-096) | In progress | Public compose contract + Watchtower hook; release versioning (US-05) outstanding                    |
-| Database Operations (PRD-060)          | Done        | Drizzle on startup, production guards, pre-migration backups                                         |
-| MCP Interface (PRD-102)                | In progress | HTTP MCP server, 14 tools across 4 domains, Docker compose profile; tests + CI image publish pending |
+| Epic                                   | Status      | Notes                                                                                                   |
+| -------------------------------------- | ----------- | ------------------------------------------------------------------------------------------------------- |
+| CI/CD workflows (PRD-016)              | Done        | Per-area quality + docker-build + publish-images.yml; no deploy step in this repo                       |
+| Application packaging & GHCR (PRD-096) | In progress | Public compose contract + Watchtower hook; release versioning (US-05) outstanding                       |
+| Database Operations (PRD-060)          | Done        | Drizzle on startup, production guards, pre-migration backups                                            |
+| MCP Interface (PRD-102)                | In progress | HTTP MCP server, 14 tools across 4 domains, Docker compose profile, CI image publish; tests in progress |
 
 #### Cortex Infrastructure
