+++ docs/themes/00-platform/prds/102-mcp-server/us-06-docker-packaging.md
@@ -9,13 +9,10 @@ Package pops-mcp as a Docker image and wire it into both compose files as an opt
 
 ## Acceptance Criteria
 
-- [x] `apps/pops-mcp/Dockerfile` — multi-stage build; builder copies shared package sources for type resolution; runtime image contains only `dist/` and production node_modules
-- [x] `infra/docker-compose.dev.yml` — `pops-mcp` service under `profiles: [mcp]`; builds from local source; uses `pops_api_key` Docker secret; depends on `pops-api` healthy
-- [x] `infra/docker-compose.yml` — `pops-mcp` service under `profiles: [mcp]`; uses `ghcr.io/knoxio/pops-mcp:${POPS_IMAGE_TAG:-main}`; `com.centurylinklabs.watchtower.enable: true`
+- [x] Multi-stage build: builder stage resolves types from shared packages; runtime image contains only compiled output and production dependencies
+- [x] Dev compose: `pops-mcp` service under `profiles: [mcp]`; builds from local source; accepts API key via Docker secret; depends on pops-api healthy
+- [x] Production compose: `pops-mcp` service under `profiles: [mcp]`; pulls from the GHCR registry; Watchtower auto-rollout enabled
 - [x] Port 3002 is bound to `${MCP_BIND_ADDR:-0.0.0.0}:3002` so local network clients can connect
 - [x] `GET /health` endpoint passes the Docker healthcheck
 - [x] `mise dev:mcp` task starts pops-mcp in dev mode (requires pops-api running separately)
-
-## Gaps (tracked)
-
-- [ ] CI publish pipeline for `ghcr.io/knoxio/pops-mcp` image — tracked as US-08
+- [x] Runtime container runs as non-root user
